requirements.txt
django>=5.0,<6.0
psycopg2-binary
django-environ
django-allauth
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-crispy-forms
crispy-bootstrap5
django-debug-toolbar
django-redis
redis
celery
django-celery-results
django-celery-beat
django-storages
boto3
django-filter
drf-spectacular
pytest-django
pytest-cov
factory-boy
pytest-mock
gunicorn
whitenoise
